refactor: step hooks & introduce escalation mode in agents
