+++ packages/pieces/community/agent/package.json
@@ -1,4 +1,4 @@
 {
   "name": "@activepieces/piece-agent",
-  "version": "0.1.1"
+  "version": "0.1.2"
 }
