+++ packages/pieces/community/agent/src/lib/actions/run-agent.ts
@@ -1,5 +1,11 @@
 import { createAction, Property, PieceAuth } from '@activepieces/pieces-framework';
-import { Agent, AgentTestResult, apId, ExecutionType, PauseType, RunAgentRequest, SeekPage, Todo } from '@activepieces/shared';
+import { Agent, AgentTestResult, AgentTaskStatus, apId, ExecutionType, PauseType, RunAgentRequest, SeekPage, Todo } from '@activepieces/shared';
+
+
+enum EscalationMode {
+  STOP = 'stop',
+  IGNORE = 'ignore',
+}
 
 export const runAgent = createAction({
   name: 'run_agent',
@@ -43,6 +49,24 @@ export const runAgent = createAction({
       description: 'Describe what you want the assistant to do.',
       required: true,
     }),
+    escalation: Property.StaticDropdown<EscalationMode>({
+      displayName: 'Escalation',
+      description: 'The behavior when the agent fails to complete the task.',
+      required: true,
+      options: {
+        options: [
+          {
+            label: 'Stop the flow',
+            value: EscalationMode.STOP,
+          },
+          {
+            label: 'Ignore and continue',
+            value: EscalationMode.IGNORE,
+          },
+        ],
+      },
+      defaultValue: EscalationMode.STOP,
+    }),
   },
   async run(context) {
     const serverToken = context.server.token;
@@ -50,10 +74,8 @@ export const runAgent = createAction({
     const agentId = context.propsValue.agentId;
 
     if (context.executionType === ExecutionType.BEGIN) {
-      const stateId = `__agent_todo_id_${apId()}`
       const actionLink = context.generateResumeUrl({
         queryParams: {
-          stateId: stateId,
         },
       });
       context.run.pause({
@@ -78,19 +100,16 @@ export const runAgent = createAction({
         body: JSON.stringify(request)
       });
       const todo = await response.json() as Todo
-      await context.store.put(stateId, todo.id)
       return {
         todoId: todo.id,
       }
     } else {
-      const todoId = await context.store.get<string>(context.resumePayload.queryParams['stateId'])
       const result = context.resumePayload.body as AgentTestResult
-      const output: AgentTestResult = {
-        todoId: todoId!,
-        output: result.output,
-        steps: result.steps,
+
+      if (result.status == AgentTaskStatus.FAILED && context.propsValue.escalation == EscalationMode.STOP) {
+        throw new Error(JSON.stringify(result))
       }
-      return output
+      return result
     }
   },
 });
