+++ packages/react-ui/src/app/builder/flow-canvas/nodes/step-node.tsx
@@ -8,6 +8,7 @@ import { useBuilderStateContext } from '@/app/builder/builder-hooks';
 import { PieceSelector } from '@/app/builder/pieces-selector';
 import { InvalidStepIcon } from '@/components/custom/alert-icon';
 import { Button } from '@/components/ui/button';
+import ImageWithFallback from '@/components/ui/image-with-fallback';
 import { LoadingSpinner } from '@/components/ui/spinner';
 import {
   Tooltip,
@@ -15,8 +16,7 @@ import {
   TooltipTrigger,
 } from '@/components/ui/tooltip';
 import { flowRunUtils } from '@/features/flow-runs/lib/flow-run-utils';
-import { PieceIcon } from '@/features/pieces/components/piece-icon';
-import { piecesHooks } from '@/features/pieces/lib/pieces-hook';
+import { stepsHooks } from '@/features/pieces/lib/steps-hooks';
 import { cn } from '@/lib/utils';
 import {
   Action,
@@ -112,7 +112,7 @@ const ApStepCanvasNode = React.memo(
     const step =
       flowStructureUtil.getStep(data.step!.name, flowVersion.trigger) ||
       data.step!;
-    const { stepMetadata } = piecesHooks.useStepMetadata({
+    const { stepMetadata } = stepsHooks.useStepMetadata({
       step,
     });
 
@@ -224,27 +224,25 @@ const ApStepCanvasNode = React.memo(
               asChild={true}
             >
               <div
-                className="flex h-full w-full"
+                className="flex items-center justify-center h-full w-full gap-3"
                 onClick={(e) => {
                   if (!openPieceSelector) {
                     handleStepClick(e);
                   }
                 }}
               >
+                <div
+                  className={cn('flex items-center justify-center h-full ', {
+                    'opacity-80': isSkipped,
+                  })}
+                >
+                  <ImageWithFallback
+                    src={stepMetadata?.logoUrl}
+                    alt={stepMetadata?.displayName}
+                    className="w-12 h-12"
+                  />
+                </div>
                 <div className="flex h-full items-center justify-between gap-3 w-full">
-                  <div className="flex items-center justify-center min-w-[46px] h-full">
-                    <div className={isSkipped ? 'opacity-80' : ''}>
-                      <PieceIcon
-                        logoUrl={
-                          step.settings?.inputUiInfo?.customizedInputs
-                            ?.logoUrl ?? stepMetadata?.logoUrl
-                        }
-                        displayName={stepMetadata?.displayName}
-                        showTooltip={false}
-                        size={'lg'}
-                      ></PieceIcon>
-                    </div>
-                  </div>
                   <div className="grow flex flex-col items-start justify-center min-w-0 w-full">
                     <div className=" flex items-center justify-between min-w-0 w-full">
                       <div
