+++ packages/react-ui/src/app/builder/index.tsx
@@ -124,7 +124,7 @@ const BuilderPage = () => {
   const leftSidePanelRef = useRef<HTMLDivElement>(null);
   const rightSidePanelRef = useRef<HTMLDivElement>(null);
 
-  const { versions, refetch: refetchPiece } =
+  const { pieceModel, refetch: refetchPiece } =
     piecesHooks.useMostRecentAndExactPieceVersion({
       name: memorizedSelectedStep?.settings.pieceName,
       version: memorizedSelectedStep?.settings.pieceVersion,
@@ -133,9 +133,6 @@ const BuilderPage = () => {
         memorizedSelectedStep?.type === TriggerType.PIECE,
     });
 
-  const pieceModel = versions
-    ? versions[memorizedSelectedStep?.settings.pieceVersion || '']
-    : undefined;
   const socket = useSocket();
 
   const { mutate: fetchAndUpdateRun } = useMutation({
