+++ packages/react-ui/src/app/builder/pieces-selector/piece-selector-utils.ts
@@ -379,19 +379,6 @@ const useAdjustPieceListHeightToAvailableSpace = (
   };
 };
 
-const isAgentPiece = (action: Action) => {
-  return (
-    action.type === ActionType.PIECE &&
-    action.settings.pieceName === '@activepieces/piece-agent'
-  );
-};
-
-const getAgentId = (action: Action) => {
-  return 'input' in action.settings && 'agentId' in action.settings.input
-    ? (action.settings.input.agentId as string)
-    : null;
-};
-
 export const pieceSelectorUtils = {
   getDefaultStep,
   isCorePiece,
@@ -405,6 +392,4 @@ export const pieceSelectorUtils = {
   isFlowController,
   isUniversalAiPiece,
   useAdjustPieceListHeightToAvailableSpace,
-  isAgentPiece,
-  getAgentId,
 };
