+++ packages/react-ui/src/app/routes/agents/builder/field-type-icon.tsx
