+++ packages/react-ui/src/app/routes/mcp-servers/id/mcp-piece-tool-dialog/index.tsx
@@ -23,7 +23,7 @@ import {
 } from '@/components/ui/tooltip';
 import { useToast } from '@/components/ui/use-toast';
 import { mcpApi } from '@/features/mcp/lib/mcp-api';
-import { piecesHooks } from '@/features/pieces/lib/pieces-hook';
+import { stepsHooks } from '@/features/pieces/lib/steps-hooks';
 import { PieceStepMetadataWithSuggestions } from '@/features/pieces/lib/types';
 import type { McpWithTools } from '@activepieces/shared';
 import { isNil, McpToolType } from '@activepieces/shared';
@@ -53,7 +53,7 @@ export function McpPieceDialog({
   const [debouncedQuery] = useDebounce(searchQuery, 300);
   const [showValidationErrors, setShowValidationErrors] = useState(false);
   const { metadata, isLoading: isPiecesLoading } =
-    piecesHooks.useAllStepsMetadata({
+    stepsHooks.useAllStepsMetadata({
       searchQuery: debouncedQuery,
       type: 'action',
     });
