+++ packages/react-ui/src/app/routes/todos/todo-details/todo-timeline-status.tsx
@@ -1,4 +1,10 @@
-import { STATUS_VARIANT, Todo } from '@activepieces/shared';
+import {
+  AGENT_REJECTED_STATUS_OPTION,
+  AGENT_RESOLVED_STATUS_OPTION,
+  isNil,
+  STATUS_VARIANT,
+  Todo,
+} from '@activepieces/shared';
 
 interface TodoTimelineStatusProps {
   todo: Todo;
@@ -11,7 +17,7 @@ export const TodoTimelineStatus = ({ todo }: TodoTimelineStatusProps) => {
 
   const bgColor = {
     [STATUS_VARIANT.POSITIVE]: 'bg-emerald-700',
-    [STATUS_VARIANT.NEGATIVE]: 'bg-rose-700',
+    [STATUS_VARIANT.NEGATIVE]: 'bg-destructive-300',
     [STATUS_VARIANT.NEUTRAL]: 'bg-gray-700',
   }[todo.status.variant];
 
@@ -20,8 +26,21 @@ export const TodoTimelineStatus = ({ todo }: TodoTimelineStatusProps) => {
       <div
         className={`${bgColor} text-background text-sm rounded-sm py-1 px-4`}
       >
-        This todo is now marked as {todo.status.name}
+        {getStatusMessage(todo)}
       </div>
     </div>
   );
 };
+
+function getStatusMessage(todo: Todo) {
+  if (!isNil(todo.agentId)) {
+    if (todo.status.name === AGENT_RESOLVED_STATUS_OPTION.name) {
+      return 'The agent has completed the task';
+    }
+    if (todo.status.name === AGENT_REJECTED_STATUS_OPTION.name) {
+      return 'The agent were unable to complete the task';
+    }
+  }
+
+  return `This todo is now marked as ${todo.status.name}`;
+}
