+++ packages/react-ui/src/features/pieces/lib/pieces-api.ts
@@ -10,51 +10,13 @@ import {
   InputPropertyMap,
 } from '@activepieces/pieces-framework';
 import {
-  Action,
-  ActionType,
   AddPieceRequestBody,
   GetPieceRequestParams,
   GetPieceRequestQuery,
   ListPiecesRequestQuery,
-  LocalesEnum,
   PackageType,
   PieceOptionRequest,
-  spreadIfDefined,
-  Trigger,
-  TriggerType,
 } from '@activepieces/shared';
-
-import { PieceStepMetadata, StepMetadata } from './types';
-
-export const CORE_STEP_METADATA: Record<
-  Exclude<ActionType, ActionType.PIECE> | TriggerType.EMPTY,
-  StepMetadata
-> = {
-  [ActionType.CODE]: {
-    displayName: t('Code'),
-    logoUrl: 'https://cdn.activepieces.com/pieces/code.svg',
-    description: t('Powerful Node.js & TypeScript code with npm'),
-    type: ActionType.CODE as const,
-  },
-  [ActionType.LOOP_ON_ITEMS]: {
-    displayName: t('Loop on Items'),
-    logoUrl: 'https://cdn.activepieces.com/pieces/loop.svg',
-    description: 'Iterate over a list of items',
-    type: ActionType.LOOP_ON_ITEMS as const,
-  },
-  [ActionType.ROUTER]: {
-    displayName: 'Router',
-    logoUrl: 'https://cdn.activepieces.com/pieces/branch.svg',
-    description: t('Split your flow into branches depending on condition(s)'),
-    type: ActionType.ROUTER,
-  },
-  [TriggerType.EMPTY]: {
-    displayName: t('Empty Trigger'),
-    logoUrl: 'https://cdn.activepieces.com/pieces/empty-trigger.svg',
-    description: t('Empty Trigger'),
-    type: TriggerType.EMPTY as const,
-  },
-};
 
 export const piecesApi = {
   list(request: ListPiecesRequestQuery): Promise<PieceMetadataModelSummary[]> {
@@ -111,82 +73,6 @@ export const piecesApi = {
             : defaultStateForDropdownProperty
         ) as ExecutePropsResult<T>;
       });
-  },
-  mapToMetadata(
-    type: 'action' | 'trigger',
-    piece: PieceMetadataModelSummary | PieceMetadataModel,
-  ): PieceStepMetadata {
-    return {
-      displayName: piece.displayName,
-      logoUrl: piece.logoUrl,
-      description: piece.description,
-      type: type === 'action' ? ActionType.PIECE : TriggerType.PIECE,
-      pieceType: piece.pieceType,
-      pieceName: piece.name,
-      pieceVersion: piece.version,
-      categories: piece.categories ?? [],
-      packageType: piece.packageType,
-      auth: piece.auth,
-    };
-  },
-  mapToSuggestions(
-    piece: PieceMetadataModelSummary,
-  ): Pick<PieceMetadataModelSummary, 'suggestedActions' | 'suggestedTriggers'> {
-    return {
-      suggestedActions: piece.suggestedActions,
-      suggestedTriggers: piece.suggestedTriggers,
-    };
-  },
-  async getMetadata(
-    step: Action | Trigger,
-    locale: LocalesEnum,
-  ): Promise<StepMetadata> {
-    const customLogoUrl =
-      'customLogoUrl' in step ? step.customLogoUrl : undefined;
-    switch (step.type) {
-      case ActionType.ROUTER:
-      case ActionType.LOOP_ON_ITEMS:
-      case ActionType.CODE:
-      case TriggerType.EMPTY:
-        return {
-          ...CORE_STEP_METADATA[step.type],
-          ...spreadIfDefined('logoUrl', customLogoUrl),
-        };
-      case ActionType.PIECE:
-      case TriggerType.PIECE: {
-        const { pieceName, pieceVersion } = step.settings;
-        const piece = await piecesApi.get({
-          name: pieceName,
-          version: pieceVersion,
-          locale: locale,
-        });
-        const metadata = await piecesApi.mapToMetadata(
-          step.type === ActionType.PIECE ? 'action' : 'trigger',
-          piece,
-        );
-        const selectedAction =
-          step.type === ActionType.PIECE
-            ? piece.actions[step.settings.actionName ?? '']
-            : null;
-
-        return {
-          ...metadata,
-          ...spreadIfDefined('logoUrl', customLogoUrl),
-          errorHandlingOptions: {
-            continueOnFailure: {
-              hide:
-                selectedAction?.errorHandlingOptions?.continueOnFailure?.hide ??
-                false,
-            },
-            retryOnFailure: {
-              hide:
-                selectedAction?.errorHandlingOptions?.retryOnFailure?.hide ??
-                false,
-            },
-          },
-        };
-      }
-    }
   },
   syncFromCloud() {
     return api.post<void>(`/v1/pieces/sync`, {});
