+++ packages/shared/src/index.ts
@@ -13,7 +13,7 @@ export * from './lib/flow-run/execution/step-output'
 export * from './lib/flows/operations'
 export * from './lib/flow-run/execution/execution-output'
 export * from './lib/flow-run/execution/step-output'
-export * from './lib/agents/agent-markdown-parser'
+export * from './lib/agents/agent-output-utils'
 export * from './lib/pieces'
 export * from './lib/store-entry/dto/store-entry-request'
 export * from './lib/webhook'
